# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Slack Automotive API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthStatus {
  isLoggedIn: boolean;
  userId?: string | null;
  displayName?: string | null;
  profilePicture?: string | null;
  workspaceName?: string | null;
  configError?: string | null;
}

export interface QrCodeResponse {
  token: string;
  loginUrl: string;
  expiresIn: number;
}

export type QrStatusResponseStatus =
  (typeof QrStatusResponseStatus)[keyof typeof QrStatusResponseStatus];

export const QrStatusResponseStatus = {
  pending: "pending",
  success: "success",
  expired: "expired",
  error: "error",
} as const;

export interface QrStatusResponse {
  status: QrStatusResponseStatus;
  errorMessage?: string | null;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ConversationType =
  (typeof ConversationType)[keyof typeof ConversationType];

export const ConversationType = {
  channel: "channel",
  dm: "dm",
  group: "group",
} as const;

export interface Participant {
  id: string;
  name: string;
  imageUrl?: string | null;
}

export interface Conversation {
  id: string;
  name: string;
  snippet?: string | null;
  unreadCount: number;
  participants: Participant[];
  imageUrl?: string | null;
  timestamp?: string | null;
  isGroup: boolean;
  type: ConversationType;
}

export interface Attachment {
  id: string;
  type: string;
  url?: string | null;
  name?: string | null;
}

export interface Message {
  id: string;
  body?: string | null;
  timestamp: string;
  senderId: string;
  senderName: string;
  isOwn: boolean;
  attachments: Attachment[];
}

export interface SendMessageRequest {
  body: string;
}

export type GetConversationsParams = {
  limit?: number;
};

export type GetMessagesParams = {
  limit?: number;
  /**
   * Pagination cursor
   */
  cursor?: string;
};