# lib/api-zod/src/generated/types/authStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Slack Automotive API
 * OpenAPI spec version: 0.2.0
 */

export interface AuthStatus {
  isLoggedIn: boolean;
  userId?: string | null;
  displayName?: string | null;
  profilePicture?: string | null;
  workspaceName?: string | null;
  configError?: string | null;
}