# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Slack Automotive API
 * OpenAPI spec version: 0.2.0
 */

export interface ErrorResponse {
  error: string;
  message: string;
}