# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Slack Automotive API
 * OpenAPI spec version: 0.2.0
 */
import type { Attachment } from "./attachment";

export interface Message {
  id: string;
  body?: string | null;
  timestamp: string;
  senderId: string;
  senderName: string;
  isOwn: boolean;
  attachments: Attachment[];
}