# lib/api-zod/src/generated/types/qrStatusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Slack Automotive API
 * OpenAPI spec version: 0.2.0
 */
import type { QrStatusResponseStatus } from "./qrStatusResponseStatus";

export interface QrStatusResponse {
  status: QrStatusResponseStatus;
  errorMessage?: string | null;
}