# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Slack Automotive API
 * OpenAPI spec version: 0.2.0
 */

export interface SuccessResponse {
  success: boolean;
  message?: string;
}